Add fuzzy scoring for contact search
